chore(swap): reorganize and cleanup unused components
