+++ ui/app/pages/swap/swap-content/swap-error-message/index.js
@@ -1,0 +1,1 @@
+export { default } from './swap-error-message.container'
