--- ui/app/pages/swap/swap-content/swap-row-wrapper/index.js
@@ -1,1 +1,0 @@
-export { default } from './swap-row-wrapper.component'
