--- ui/app/pages/swap/swap-content/swap-row-wrapper/swap-row-error-message/index.js
@@ -1,1 +1,0 @@
-export { default } from './swap-row-error-message.container'
